requirements.txt
# 核心依赖
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# 计算机视觉
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 配置文件
PyYAML>=6.0
PyQt5>=5.15.0

# 其他工具
tqdm>=4.65.0
scikit-learn>=1.3.0
